rename the unified module to `component`
